requirements.txt
pandas
seaborn
streamlit
plotly.express